Div0 Talk on 15 January 2019
Talk abstract: With the increasing popularisation of cloud computing in the 2010s, an increasing amount of attention is being paid by simulated "adversaries" to remote social engineering and technical techniques. Quite a bit of attention has shifted away from protection from the client end to the provider end of the relationship. This talk uses analogies - bringing participants from the typical technical IT security controls & standards to identify physical security controls in the real world. Also, familiar security models in the IT security space such as STRIDE, surface area reduction, etc. will be shown to work in the physical world. It is hoped that with this type of thought process, security professionals will be able to identify new weaknesses in the physical world that attackers may use in exploiting information systems.